Add test command and include in pr flow [minor] (#33)
This PR adds a new command ("gh dxp test") that will look for one of
four possible files and run a test command based on these. If no command
is found, it will proceed. If a command is found, but returns an error,
it will fail.

The command is also automatically called in the PR flow.

This does _not_ introduce configuration of the test command using the
.devxp file

Summary:
- add test command


Testing:
- [x] Unit Tests
- [ ] Integration Tests
- Test Command: make test


Documentation:
- No updates