handle file renames [patch] (#74)
Summary:
when using 'git mv' to rename files (and perhaps other situations as
well), git status will show the file as renamed as opposed to the more
usual "old file deleted, new file created". This caused a bug where the
pr command was trying to git add a string on the form "oldname.txt ->
newname.txt". This change parses these lines so that the new name is
included rather than the entire string.

Testing:
- [x] Unit Tests
- [ ] Integration Tests


Documentation:
- No updates