Update the PR command handling (#78)
Changes the command structure for handling Pull Requests. The new
structure basically has subcommands of dxp pr:
* dxp pr create
* dxp pr merge
* dxp pr update

We do not create subcommands for things that the gh pr commands already
cover. (e.g. list) That could be done, but I think that is an
anti-pattern. My thinking is that we only do this if we see a need to
expand upon functionality.