Refactor lint_test.go (#101)
## 📝 Description

Since the main required behaviour of the lint command is to call
mega-linter-runner with the correct arguments, we need to
AssertExpectations (i.e. check that all mocks have been run). This
didn't mesh well with creating test cases in a for loop as we've done
elsewhere, because it was leading to lots of logic and conditional
mocking. This diff refactors the tests into separate functions.

## 📋 Checklist

* ✅ Lint checks passed on local machine.
* ✅ Unit tests passed on local machine.
* ✅ This PR adds news tests.