Replace explicit file inclusion in git add with -u [patch]
Some users have encountered a bug in which the command _git add x_ will fail if x is a deleted file _and_ it is already staged. I think this behaviour is somewhat strange. Nonetheless, I've replaced the git add command that has explicit file inclusions with -u, which should add all tracked files. This seems to resolve the issue, and is also a neater solution overall, in my opinion.

## 📋 Checklist

* ⛔ **This PR has not been linted! The --nolint option was used.**
* ✅ Unit tests passed on local machine.