Add clone-all functionality to gh-dxp
## 📝 Description

This adds a new clone-all command. Due to issues we have noted with the API,
this does not try to parallelize clones; everything is done sequentially to keep
things simple.

Currently this grabs everything that the user has access to. An --owner or--organization
option should probably be introduced to allow the user to limit which organization
repos are cloned.

## 🔗 Issue ID(s): TDX-741

## 📋 Checklist

* ✅ Lint checks passed on local machine.
* ✅ Unit tests passed on local machine.
* ✅ This PR adds new tests.