Introduce retry logic to clone operation (#121)
## 📝 Description

When cloning all or most of the repos in the Elhub org, several users
have reported getting 'connection reset' error messages for some of the
repos. This could be some kind of throttling by github - necessitating
rerunning the script at least once to get all the repos.
This diff introduces a one second delay to each clone, which does seem
to prevent this issue _on my machine_ but also slows the total clone
time down by a lot.

**EDIT: I've replaced the delay with some retry logic that seems to do
the trick**

## 📋 Checklist

* ✅ Lint checks passed on local machine.
* ✅ Unit tests passed on local machine.